[DATALAD RUNCMD] Fix (in text and rename folder) 'ancilliary'
=== Do not change lines below ===
{
 "chain": [],
 "cmd": "git mv ancilliary ancillary; codespell -w ./tests/utils/manifest_tester.py ./tests/test_shextest_validation/test_manifest_shex_shexc.py",
 "exit": 0,
 "extra_inputs": [],
 "inputs": [],
 "outputs": [],
 "pwd": "."
}
^^^ Do not change lines above ^^^
