+++ ancillary/earlreport.py
